Fix run object
