+++ src/v1/run.rs
@@ -3,6 +3,7 @@ use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use std::collections::HashMap;
 
+use super::types;
 use crate::impl_builder_methods;
 
 #[derive(Debug, Serialize, Clone)]
@@ -95,7 +96,7 @@ pub struct RunObject {
     pub completed_at: Option<i64>,
     pub model: String,
     pub instructions: Option<String>,
-    pub tools: Vec<HashMap<String, String>>,
+    pub tools: Vec<types::Tools>,
     pub metadata: HashMap<String, String>,
     pub headers: Option<HashMap<String, String>>,
 }
